file naming changes, more styling changes
