+++ src/components/CodingRoles.jsx
@@ -4,9 +4,9 @@ import {
   genomicsDevOps,
   genomicsIT, mondrianIT,
   starlingSoftwareEngineer
-} from '../CvContents/RolesContents';
+} from '../cvcontents/rolesContents';
 
-export default function CodingRoles ({ chipClass }) {
+export default function CodingRoles () {
   return (
     <div>
       <RoleSection
@@ -15,7 +15,7 @@ export default function CodingRoles ({ chipClass }) {
         yearRange='Jan 2018 - Jul 2020'
         jobTitle='Software Engineer'
         company='Starling Bank'
-        chipClass={chipClass}
+        chipClass='-coding'
       />
       <RoleSection
         year='2017'
@@ -23,7 +23,7 @@ export default function CodingRoles ({ chipClass }) {
         yearRange='July 2016 - Nov 2017'
         jobTitle='DevOps Engineer'
         company='Genomics England'
-        chipClass={chipClass}
+        chipClass='-coding'
       />
       <RoleSection
         year='2016'
@@ -31,7 +31,7 @@ export default function CodingRoles ({ chipClass }) {
         yearRange='Nov 2015 - July 2016'
         jobTitle='IT Analyst'
         company='Genomics England'
-        chipClass={chipClass}
+        chipClass='-coding'
       />
       <RoleSection
         year='2015'
@@ -39,7 +39,7 @@ export default function CodingRoles ({ chipClass }) {
         yearRange='Sept 2015 - Nov 2016'
         jobTitle='IT Assistant'
         company='Mondrian London'
-        chipClass={chipClass}
+        chipClass='-coding'
       />
     </div>
   )
