+++ src/components/CvBody.scss
@@ -1,0 +1,20 @@
+@import 'colours';
+
+.app-section-container {
+  display: grid;
+  grid-template-columns: 1fr 15rem;
+}
+
+.app-switcher-bar {
+  align-self: center;
+  grid-column-start: 2;
+  margin: 10px 0;
+
+  &-coding {
+    color: $red;
+  }
+
+  &-pre-coding {
+    color: $purple;
+  }
+}
