--- src/components/_colours.scss
@@ -1,2 +1,0 @@
-$red: #f50057;
-$purple: #7b08c3;
