+++ src/components/colours.scss
@@ -1,0 +1,3 @@
+$red: #f50057;
+$purple: #7b08c3;
+$text: #4d4545;
