+++ src/cvcontents/rolesContents.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Typography } from '@material-ui/core';
 
 export const starlingSoftwareEngineer =
-  <Typography variant='body2' style={{ padding: '3px'}}>
+  <Typography variant='body2'>
     • Full Stack Engineer working principally with Java/Postgres back end, and React front end<br/>
     • Worked in the Marketplace and Public APIs team, spanning the Starling 3rd party integrations marketplace, Developer Sandbox, public API and Open Banking<br/>
     • Was part of the small team that delivered the Starling personal and business Euro currency accounts and SEPA payment scheme.<br/>
@@ -10,7 +10,7 @@ export const starlingSoftwareEngineer =
   </Typography>;
 
 export const genomicsDevOps =
-  <Typography variant='body2' style={{ padding: '3px'}}>
+  <Typography variant='body2'>
    • Internally promoted after demonstrating an interest and aptitude for scripting and automation.<br/>
    • Responsibilities revolved around aiding Genomics England in delivery of its infrastructure and platforms.<br/>
    • Working with developers in a time-critical environment to deploy bespoke applications on our infrastructure, and to automate future deployments.<br/>
@@ -18,7 +18,7 @@ export const genomicsDevOps =
   </Typography>;
 
 export const genomicsIT =
-  <Typography variant='body2' style={{ padding: '3px'}}>
+  <Typography variant='body2'>
     • Assumed control over internal IT support and systems administration.
     These responsibilities had previously been divided and handled by non-specialists and
     external support.<br/>
@@ -31,14 +31,14 @@ export const genomicsIT =
   </Typography>;
 
 export const mondrianIT =
-  <Typography variant='body2' style={{ padding: '3px'}}>
+  <Typography variant='body2'>
     • Structured training role, reporting to Director of IT<br/>
     • Responsible for resolution of day to day IT related queries and problems via service desk
     management system.
   </Typography>;
 
 export const uniqueVoice =
-  <Typography variant='body2' style={{ padding: '3px'}}>
+  <Typography variant='body2'>
     • Learned and performed roles on stage or film as directed.<br/>
     • Interpreted and memorised lines on provided scripts, including directorial notes
     and actions.<br/>
@@ -47,7 +47,7 @@ export const uniqueVoice =
   </Typography>;
 
 export const monarchRecruitment =
-  <Typography variant='body2' style={{ padding: '3px'}}>
+  <Typography variant='body2'>
     • Used current IT in every lesson in order to promote up to date skills and creative
     teaching.<br/>
     • Was able to adapt and improvise to quickly changing circumstances and variable
@@ -55,14 +55,14 @@ export const monarchRecruitment =
   </Typography>;
 
 export const universityWestEngland =
-  <Typography variant='body2' style={{ padding: '3px'}}>
+  <Typography variant='body2'>
     • PGCE PCET (Post-Compulsory Education and Training)<br/>
     • Qualified to teach Performing Arts disciplines at post-GCSE level<br/>
     • Passed at Masters Level
   </Typography>;
 
 export const aldSeniorCredit =
-  <Typography variant='body2' style={{ padding: '3px'}}>
+  <Typography variant='body2'>
     • Responsibilities included ownership of the Synertec Prism mailing system for the depart ment.
     This included 1st line support for users on the team, staff training and contact with Synertec. <br/>
     • Used in-house software and Excel to generate regular reports using V-Lookups and Pivot
@@ -73,6 +73,6 @@ export const aldSeniorCredit =
   </Typography>;
 
 export const deMontfort =
-  <Typography variant='body2' style={{ padding: '3px'}}>
+  <Typography variant='body2'>
     • Graduated with 2:1 Honours
   </Typography>
