+++ src/cvcontents/sectionContents.jsx
@@ -2,10 +2,9 @@ import React from 'react';
 import Breadcrumbs from '@material-ui/core/Breadcrumbs';
 import Typography from '@material-ui/core/Typography';
 import TripOrigin from '@material-ui/icons/TripOrigin';
-import { red, purple } from "../components/colors";
 
 export const personalProfileContents =
-  <Typography color='secondary' > Creative, proactive and engaging with an inquisitive nature and desire to learn. Has a
+  <Typography> Creative, proactive and engaging with an inquisitive nature and desire to learn. Has a
   background in various disciplines, both creative and technical, resulting in a broad skill-set
   and a unique approach to problem-solving. Approaches work with a desire to achieve and
   excel. Has had a life-long obsession with technology and continues to pursue this through
@@ -31,7 +30,7 @@ export const skills = [
 ];
 
 export const skillsContents =
-  <Breadcrumbs separator={<TripOrigin style={{ color: red }} fontSize='small'/>} maxItems={20} itemsAfterCollapse={4} itemsBeforeCollapse={4}>
+  <Breadcrumbs className='skills' separator={<TripOrigin fontSize='small'/>} maxItems={20} itemsAfterCollapse={4} itemsBeforeCollapse={4}>
     { skills.map((skill, i) => <Typography key={`skill-${i}`}>{skill}</Typography>)}
   </Breadcrumbs>;
 
@@ -50,6 +49,6 @@ const interests = [
 ];
 
 export const interestContents =
-  <Breadcrumbs separator={<TripOrigin style={{ color: purple }} fontSize='small'/>} maxItems={20} itemsAfterCollapse={4} itemsBeforeCollapse={4}>
+  <Breadcrumbs className='interests' separator={<TripOrigin fontSize='small'/>} maxItems={20} itemsAfterCollapse={4} itemsBeforeCollapse={4}>
     { interests.map((interest, i) => <Typography key={`interest-${i}`}>{interest}</Typography>)}
   </Breadcrumbs>;
